Added Shader Group to Scene graph structure
Scene graph hierarchy now has ShaderGroup which takes a shader when it
is created and applies the shader to all of its children upon a call to
draw.  The scene should be able to support multiple shaders although
that is not yet tested. Commit also includes simple test of shader by
creating a second child sphere and painting it blue.  (Also removed some
unnecessary includes)