Bezier Patches (alpha)
I have started the INITIAL workings for bezier patches. Trust me it is
not pretty at the moment and I probably need to ask Michael for tips on
how to design it prettier. But it works! It's an order 4 bezier patch
and all you need to do is provide all 16 control points for it to work!
Yay!  I've added one to the test so you can play around with the control
points if you like. It's a Geode so it can go into the scene graphs.
TODO: Make it easier to create and modify them.